Major refactoring to be extracted into different PRs
