+++ src/DataGenerator/CharCountBased.php
@@ -3,37 +3,17 @@ declare(strict_types=1);
 
 namespace FastRoute\DataGenerator;
 
-use function count;
-use function implode;
-
+/**
+ * @deprecated Pass the chunk processor to the constructor of your data generator instead
+ */
 class CharCountBased extends RegexBasedAbstract
 {
-    protected function getApproxChunkSize(): int
+    protected function getChunkProcessor(): ChunkProcessorInterface
     {
-        return 30;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    protected function processChunk(array $regexToRoutesMap): array
-    {
-        $routeMap = [];
-        $regexes = [];
-
-        $suffixLen = 0;
-        $suffix = '';
-        $count = count($regexToRoutesMap);
-        foreach ($regexToRoutesMap as $regex => $route) {
-            $suffixLen++;
-            $suffix .= "\t";
-
-            $regexes[] = '(?:' . $regex . '/(\t{' . $suffixLen . '})\t{' . ($count - $suffixLen) . '})';
-            $routeMap[$suffix] = [$route->handler, $route->variables, $route];
+        if ($this->chunkProcessor === null) {
+            $this->chunkProcessor = new CharCountProcessor();
         }
 
-        $regex = '~^(?|' . implode('|', $regexes) . ')$~';
-
-        return ['regex' => $regex, 'suffix' => '/' . $suffix, 'routeMap' => $routeMap];
+        return $this->chunkProcessor;
     }
 }
