+++ src/Dispatcher.php
@@ -3,23 +3,11 @@ declare(strict_types=1);
 
 namespace FastRoute;
 
-use FastRoute\Dispatcher\Result;
+use FastRoute\Dispatcher\DispatcherInterface;
 
-interface Dispatcher
+/**
+ * Dispatcher
+ */
+interface Dispatcher extends DispatcherInterface
 {
-    public const NOT_FOUND = 0;
-    public const FOUND = 1;
-    public const METHOD_NOT_ALLOWED = 2;
-
-    /**
-     * Dispatches against the provided HTTP method verb and URI.
-     *
-     * Returns array with one of the following formats:
-     *
-     *     [self::NOT_FOUND]
-     *     [self::METHOD_NOT_ALLOWED, ['GET', 'OTHER_ALLOWED_METHODS']]
-     *     [self::FOUND, $handler, ['varName' => 'value', ...]]
-     *
-     */
-    public function dispatch(string $httpMethod, string $uri): Result;
 }
